Program Organization Initial Proposal
This document is a proposal for organizing all the code we worked with and tested into a set of main python scripts users could execute from the command line. The first idea is to have one responsible for downloading (if needed), copying, and potentially formatting files before later processing. The second set includes one program for Optical Flow execution and one for Tiramisu execution. These main programs will take responsibility for reading command line parameters and passing appropriate values to the classes that are responsible for these operations. As of now this document isn't complete, but work will continue tonight and into tomorrow...